refactor(AI): implement IEnemy interface

+++ Assets/Script/AI/Gator/FireBallAttack.cs
@@ -4,7 +4,7 @@ namespace Script.AI.Gator
 {
     public class FireBallAttack : HitBox.Attack
     {
-        public override float Damage(GameObject target)
+        public override int Damage(GameObject target)
         {
             transform.GetComponent<FireBall>().OnCollide();
             return base.Damage(target);
